feat(notifications): Configure OneSignal custom link
Integrate OneSignal's custom link feature for a more controlled and on-brand notification subscription experience. This change disables the default OneSignal notification button and enables a custom button within the application's UI.

Additionally, this commit includes styling adjustments to ensure the custom notification button is properly displayed and a fallback mechanism to manually request permissions if the OneSignal button fails to load.